backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ultralytics
opencv-python
aiofiles
python-dotenv
google-cloud-aiplatform
Pillow
